basic folder implementation, change save files to json
