+++ src/main/java/one/pouekdev/coordinatelist/CListData.java
@@ -1,34 +1,61 @@
 package one.pouekdev.coordinatelist;
 
 import com.google.common.collect.Lists;
+import com.google.gson.Gson;
+import com.google.gson.reflect.TypeToken;
 import net.fabricmc.loader.api.FabricLoader;
 
 import java.io.*;
+import java.lang.reflect.Type;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
+import java.nio.file.Path;
 import java.util.List;
 
 public class CListData{
-    public static void saveListToFile(String fileName, List<CListWaypoint> waypointList){
-        if(!Files.exists(FabricLoader.getInstance().getConfigDir().resolve("coordinatelist"))){
+    private static final Path modDir = FabricLoader.getInstance().getConfigDir().resolve("coordinatelist");
+    private static final File modDirFile = modDir.toFile();
+
+    public static void saveListToFile(String fileName, CListElementsContainer dataContainer){
+        if(!Files.exists(modDir)){
             try{
-                Files.createDirectories(FabricLoader.getInstance().getConfigDir().resolve("coordinatelist"));
+                Files.createDirectories(modDir);
             }
             catch(IOException ignored){}
         }
-        File dataDir = FabricLoader.getInstance().getConfigDir().resolve("coordinatelist").toFile();
-        File file = new File(dataDir, fileName);
+        File file = new File(modDirFile, fileName);
         try(PrintWriter writer = new PrintWriter(new BufferedWriter(new OutputStreamWriter(new FileOutputStream(file), StandardCharsets.UTF_8)))){
-            for(int i = 0; i < waypointList.size(); i++){
-                writer.println(CListClient.variables.waypoints.get(i).getCoordinates() + "~" + CListClient.variables.waypoints.get(i).name.replaceAll("~", "") + "~" + CListClient.variables.waypoints.get(i).dimension + "~" + CListClient.variables.waypoints.get(i).color.getHexNoAlpha() + "~" + CListClient.variables.waypoints.get(i).render + "~" + CListClient.variables.waypoints.get(i).deathpoint);
-            }
+            Gson gson = new Gson();
+            String json = gson.toJson(dataContainer);
+            writer.write(json);
         }
         catch(IOException ignored){}
     }
 
-    public static List<CListWaypoint> loadListFromFile(String fileName){
-        File dataDir = FabricLoader.getInstance().getConfigDir().resolve("coordinatelist").toFile();
-        File file = new File(dataDir, fileName);
+    public static CListElementsContainer loadListFromFile(String fileName){
+        File file = new File(modDirFile, fileName);
+        if(!file.exists()){
+            return null;
+        }
+        try(BufferedReader reader = new BufferedReader(new InputStreamReader(new FileInputStream(file), StandardCharsets.UTF_8))){
+            String json = reader.readAllAsString();
+            Gson gson = new Gson();
+            Type listType = new TypeToken<CListElementsContainer>(){}.getType();
+            return gson.fromJson(json, listType);
+        }
+        catch(IOException ignored){}
+        return null;
+    }
+
+    public static void deleteLegacyFile(String fileName){
+        File file = new File(modDirFile, fileName);
+        if(file.exists()){
+            boolean ignored = file.delete();
+        }
+    }
+
+    public static List<CListWaypoint> loadListFromFileLegacy(String fileName){
+        File file = new File(modDirFile, fileName);
         if(!file.exists()){
             return null;
         }
@@ -50,43 +77,17 @@ public class CListData{
                     catch(IndexOutOfBoundsException ignored){}
                     CListWaypoint waypoint;
                     if(color == null){
-                        waypoint = new CListWaypoint(coords, name, dimension, new CListWaypointColor(), Boolean.parseBoolean(bool), Boolean.parseBoolean(deathpoint));
+                        waypoint = new CListWaypoint(coords, name, dimension, new CListElementColor(), Boolean.parseBoolean(bool), Boolean.parseBoolean(deathpoint));
                     }
                     else{
-                        CListWaypointColor color_class = new CListWaypointColor(0, 0, 0);
-                        color_class.set(color);
-                        waypoint = new CListWaypoint(coords, name, dimension, color_class, Boolean.parseBoolean(bool), Boolean.parseBoolean(deathpoint));
+                        CListElementColor colorClass = new CListElementColor(0, 0, 0);
+                        colorClass.set(color);
+                        waypoint = new CListWaypoint(coords, name, dimension, colorClass, Boolean.parseBoolean(bool), Boolean.parseBoolean(deathpoint));
                     }
                     waypointList.add(waypoint);
                 }
             }
             return waypointList;
-        }
-        catch(IOException ignored){}
-        return null;
-    }
-
-    public static void deleteLegacyFile(String fileName){
-        File dataDir = FabricLoader.getInstance().getConfigDir().resolve("coordinatelist").toFile();
-        File file = new File(dataDir, fileName);
-        if(file.exists()){
-            boolean ignored = file.delete();
-        }
-    }
-
-    public static List<String> loadListFromFileLegacy(String fileName){
-        File dataDir = FabricLoader.getInstance().getConfigDir().resolve("coordinatelist").toFile();
-        File file = new File(dataDir, fileName);
-        if(!file.exists()){
-            return null;
-        }
-        try(BufferedReader reader = new BufferedReader(new InputStreamReader(new FileInputStream(file), StandardCharsets.UTF_8))){
-            List<String> stringList = Lists.newArrayList();
-            String line;
-            while((line = reader.readLine()) != null){
-                stringList.add(line);
-            }
-            return stringList;
         }
         catch(IOException ignored){}
         return null;
