+++ src/main/java/one/pouekdev/coordinatelist/CListWaypoint.java
@@ -1,19 +1,11 @@
 package one.pouekdev.coordinatelist;
 
-import net.minecraft.network.chat.Component;
-import org.apache.commons.lang3.StringUtils;
-
-public class CListWaypoint{
+public class CListWaypoint extends CListElement{
     public int x, y, z;
-    public String name;
-    public String dimension;
-    public CListWaypointColor color;
-    public boolean render;
     public boolean deathpoint;
     public boolean locked;
-    private int bugFix;
 
-    CListWaypoint(String coords, String waypointName, String waypointDimension, CListWaypointColor waypointColor, boolean isRendered, boolean isDeathpoint){
+    CListWaypoint(String coords, String waypointName, String waypointDimension, CListElementColor waypointColor, boolean isRendered, boolean isDeathpoint){
         String s = coords;
         s = s.replace("X", "");
         s = s.replace("Y", "");
@@ -29,10 +21,9 @@ public class CListWaypoint{
         this.render = isRendered;
         this.deathpoint = isDeathpoint;
         this.locked = false;
-        this.bugFix = 0;
     }
 
-    CListWaypoint(int x, int y, int z, String waypointName, String waypointDimension, CListWaypointColor waypointColor, boolean isRendered, boolean isDeathpoint){
+    CListWaypoint(int x, int y, int z, String waypointName, String waypointDimension, CListElementColor waypointColor, boolean isRendered, boolean isDeathpoint){
         this.x = x;
         this.y = y;
         this.z = z;
@@ -42,32 +33,5 @@ public class CListWaypoint{
         this.render = isRendered;
         this.deathpoint = isDeathpoint;
         this.locked = false;
-        this.bugFix = 0;
-    }
-
-    public String getCoordinates(){
-        return "X: " + x + " Y: " + y + " Z: " + z;
-    }
-
-    public void toggleVisibility(){
-        this.bugFix += 1;
-        if(bugFix == 2){
-            this.bugFix = 0;
-            this.render = !this.render;
-            CListClient.variables.savedSinceLastUpdate = false;
-        }
-    }
-
-    public Component getDimensionText(){
-        return Component.literal(this.getDimensionString());
-    }
-
-    public String getDimensionString(){
-        String s = this.dimension;
-        s = s.replace("minecraft:", "");
-        s = s.replace("_", " ");
-        s = s.replace(":", " ");
-        s = StringUtils.capitalize(s);
-        return s;
     }
 }
